Переименовал Extensions.cs в VkFoafExtensions.cs
